Move inlined functions to source file
