+++ source/state_representation/include/state_representation/geometry/Ellipsoid.hpp
@@ -1,10 +1,7 @@
 #pragma once
 
 #include <list>
-#include <eigen3/Eigen/Core>
-#include <eigen3/Eigen/Dense>
-#include <eigen3/Eigen/Sparse>
-#include <random>
+
 #include "state_representation/geometry/Shape.hpp"
 #include "state_representation/space/cartesian/CartesianPose.hpp"
 
@@ -13,10 +10,6 @@ namespace state_representation {
  * @class Ellipsoid
  */
 class Ellipsoid : public Shape {
-private:
-  std::vector<double> axis_lengths_; ///< axis lengths in x,y directions
-  double rotation_angle_; ///< angle of rotation around z axis of the reference frame
-
 public:
   /**
    * @brief Empty constructor
@@ -71,6 +64,17 @@ public:
   double get_axis_length(unsigned int index) const;
 
   /**
+   * @brief Getter of the rotation angle
+   * @return The rotation angle
+   */
+  double get_rotation_angle() const;
+
+  /**
+   * @brief Getter of the rotation as a pose
+   */
+  const CartesianPose get_rotation() const;
+
+  /**
    * @brief Setter of the axis lengths
    * @param axis_lengths The new values
    */
@@ -84,53 +88,10 @@ public:
   void set_axis_lengths(unsigned int index, double axis_length);
 
   /**
-   * @brief Getter of the rotation angle
-   * @return The rotation angle
-   */
-  double get_rotation_angle() const;
-
-  /**
    * @brief Setter of the rotation angle
    * @param rotation_angle The rotation angle
    */
   void set_rotation_angle(double rotation_angle);
-
-  /**
-   * @brief Getter of the rotation as a pose
-   */
-  const CartesianPose get_rotation() const;
-
-  /**
-   * @brief Function to sample an obstacle from its parameterization
-   * @param nb_samples The number of sample points to generate
-   * @return The list of sample points
-   */
-  const std::list<CartesianPose> sample_from_parameterization(unsigned int nb_samples) const;
-
-  /**
-   * @brief Compute an Ellipsoid from its algebraic equation ax2 + bxy + cy2 + cx + ey + f
-   * @return The Ellipsoid in its geometric representation
-   */
-  static const Ellipsoid from_algebraic_equation(
-      const std::string& name, const std::vector<double>& coefficients, const std::string& reference_frame = "world"
-  );
-
-  /**
-   * @brief Fit an Ellipsoid on a set of points
-   * This method uses direct least square fitting from
-   * Fitzgibbon, A., et al. (1999). "Direct least square fitting of ellipses."
-    * IEEE Transactions on pattern analysis and machine intelligence 21(5)
-   */
-  static const Ellipsoid fit(
-      const std::string& name, const std::list<CartesianPose>& points, const std::string& reference_frame = "world",
-      double noise_level = 0.01
-  );
-
-  /**
-   * @brief Convert the Ellipsoid to an std vector representation of its parameter
-   * @return An std vector with [center_position, rotation_angle, axis_lengths]
-   */
-  const std::vector<double> to_std_vector() const;
 
   /**
    * @brief Set the ellipsoid data from an Eigen vector
@@ -145,12 +106,48 @@ public:
   void set_data(const std::vector<double>& data) override;
 
   /**
+   * @brief Fit an Ellipsoid on a set of points
+   * This method uses direct least square fitting from
+   * Fitzgibbon, A., et al. (1999). "Direct least square fitting of ellipses."
+    * IEEE Transactions on pattern analysis and machine intelligence 21(5)
+   */
+  static const Ellipsoid fit(
+      const std::string& name, const std::list<CartesianPose>& points, const std::string& reference_frame = "world",
+      double noise_level = 0.01
+  );
+
+  /**
+   * @brief Compute an Ellipsoid from its algebraic equation ax2 + bxy + cy2 + cx + ey + f
+   * @return The Ellipsoid in its geometric representation
+   */
+  static const Ellipsoid from_algebraic_equation(
+      const std::string& name, const std::vector<double>& coefficients, const std::string& reference_frame = "world"
+  );
+
+  /**
+   * @brief Function to sample an obstacle from its parameterization
+   * @param nb_samples The number of sample points to generate
+   * @return The list of sample points
+   */
+  const std::list<CartesianPose> sample_from_parameterization(unsigned int nb_samples) const;
+
+  /**
+   * @brief Convert the Ellipsoid to an std vector representation of its parameter
+   * @return An std vector with [center_position, rotation_angle, axis_lengths]
+   */
+  const std::vector<double> to_std_vector() const;
+
+  /**
     * @brief Overload the ostream operator for printing
     * @param os The ostream to append the string representing the state
     * @param state The state to print
     * @return The appended ostream
      */
   friend std::ostream& operator<<(std::ostream& os, const Ellipsoid& ellipsoid);
+
+private:
+  std::vector<double> axis_lengths_; ///< axis lengths in x,y directions
+  double rotation_angle_; ///< angle of rotation around z axis of the reference frame
 };
 
 inline void swap(Ellipsoid& state1, Ellipsoid& state2) {
@@ -158,64 +155,4 @@ inline void swap(Ellipsoid& state1, Ellipsoid& state2) {
   std::swap(state1.axis_lengths_, state2.axis_lengths_);
   std::swap(state1.rotation_angle_, state2.rotation_angle_);
 }
-
-inline Ellipsoid& Ellipsoid::operator=(const Ellipsoid& state) {
-  Ellipsoid tmp(state);
-  swap(*this, tmp);
-  return *this;
-}
-
-inline const std::vector<double>& Ellipsoid::get_axis_lengths() const {
-  return this->axis_lengths_;
-}
-
-inline double Ellipsoid::get_axis_length(unsigned int index) const {
-  return this->axis_lengths_[index];
-}
-
-inline void Ellipsoid::set_axis_lengths(const std::vector<double>& axis_lengths) {
-  this->axis_lengths_ = axis_lengths;
-  this->set_empty(false);
-}
-
-inline void Ellipsoid::set_axis_lengths(unsigned int index, double axis_length) {
-  this->axis_lengths_[index] = axis_length;
-  this->set_empty(false);
-}
-
-inline double Ellipsoid::get_rotation_angle() const {
-  return this->rotation_angle_;
-}
-
-inline void Ellipsoid::set_rotation_angle(double rotation_angle) {
-  this->rotation_angle_ = rotation_angle;
-  this->set_empty(false);
-}
-
-inline const std::vector<double> Ellipsoid::to_std_vector() const {
-  if (this->is_empty()) {
-    throw exceptions::EmptyStateException(this->get_name() + " state is empty");
-  }
-  std::vector<double> representation(6);
-  // position
-  representation[0] = this->get_center_position()(0);
-  representation[1] = this->get_center_position()(1);
-  representation[2] = this->get_center_position()(2);
-  // rotation angle
-  representation[3] = this->get_rotation_angle();
-  // axis lengths
-  representation[4] = this->get_axis_length(0);
-  representation[5] = this->get_axis_length(1);
-  return representation;
-}
-
-inline const CartesianPose Ellipsoid::get_rotation() const {
-  if (this->is_empty()) {
-    throw exceptions::EmptyStateException(this->get_name() + " state is empty");
-  }
-  Eigen::Quaterniond rotation(Eigen::AngleAxisd(this->rotation_angle_, Eigen::Vector3d::UnitZ()));
-  return CartesianPose(
-      this->get_center_pose().get_name() + "_rotated", Eigen::Vector3d::Zero(), rotation,
-      this->get_center_pose().get_name());
-}
-}
+}// namespace state_representation
