+++ source/state_representation/include/state_representation/geometry/Shape.hpp
@@ -1,13 +1,9 @@
 #pragma once
 
-#include <eigen3/Eigen/Core>
 #include "state_representation/State.hpp"
 #include "state_representation/space/cartesian/CartesianState.hpp"
 #include "state_representation/space/cartesian/CartesianPose.hpp"
 #include "state_representation/space/cartesian/CartesianTwist.hpp"
-#include "state_representation/exceptions/EmptyStateException.hpp"
-#include "state_representation/exceptions/IncompatibleStatesException.hpp"
-#include "state_representation/exceptions/IncompatibleReferenceFramesException.hpp"
 
 namespace state_representation {
 /**
@@ -130,61 +126,4 @@ inline void swap(Shape& state1, Shape& state2) {
   swap(static_cast<State&>(state1), static_cast<State&>(state2));
   std::swap(state1.center_state_, state2.center_state_);
 }
-
-inline Shape& Shape::operator=(const Shape& state) {
-  Shape tmp(state);
-  swap(*this, tmp);
-  return *this;
-}
-
-inline const CartesianState& Shape::get_center_state() const {
-  return this->center_state_;
-}
-
-inline const CartesianPose& Shape::get_center_pose() const {
-  return static_cast<const CartesianPose&>(this->center_state_);
-}
-
-inline const Eigen::Vector3d Shape::get_center_position() const {
-  return this->center_state_.get_position();
-}
-
-inline const Eigen::Quaterniond Shape::get_center_orientation() const {
-  return this->center_state_.get_orientation();
-}
-
-inline const CartesianTwist& Shape::get_center_twist() const {
-  return static_cast<const CartesianTwist&>(this->center_state_);
-}
-
-inline void Shape::set_center_state(const CartesianState& state) {
-  if (state.is_empty()) {
-    throw exceptions::EmptyStateException(state.get_name() + " state is empty");
-  }
-  this->center_state_ = state;
-  this->set_empty(false);
-  this->reset_timestamp();
-}
-
-inline void Shape::set_center_pose(const CartesianPose& pose) {
-  if (this->center_state_.get_reference_frame() != pose.get_reference_frame()) {
-    throw exceptions::IncompatibleReferenceFramesException(
-        "The shape state and the given pose are not expressed in the same reference frame");
-  }
-  this->center_state_.set_pose(pose.get_position(), pose.get_orientation());
-  this->set_empty(false);
-  this->reset_timestamp();
-}
-
-inline void Shape::set_center_position(const Eigen::Vector3d& position) {
-  this->center_state_.set_position(position);
-  this->set_empty(false);
-  this->reset_timestamp();
-}
-
-inline void Shape::set_center_orientation(const Eigen::Quaterniond& orientation) {
-  this->center_state_.set_orientation(orientation);
-  this->set_empty(false);
-  this->reset_timestamp();
-}
-}
+}// namespace state_representation
