+++ source/state_representation/src/geometry/Shape.cpp
@@ -1,5 +1,8 @@
 #include "state_representation/geometry/Shape.hpp"
 
+#include "state_representation/exceptions/EmptyStateException.hpp"
+#include "state_representation/exceptions/IncompatibleReferenceFramesException.hpp"
+
 namespace state_representation {
 
 Shape::Shape() : State(), center_state_(CartesianState::Identity("")) {
@@ -23,6 +26,63 @@ Shape Shape::Unit(const std::string& name, const std::string& reference_frame) {
   return unit;
 }
 
+Shape& Shape::operator=(const Shape& state) {
+  Shape tmp(state);
+  swap(*this, tmp);
+  return *this;
+}
+
+const CartesianState& Shape::get_center_state() const {
+  return this->center_state_;
+}
+
+const CartesianPose& Shape::get_center_pose() const {
+  return static_cast<const CartesianPose&>(this->center_state_);
+}
+
+const Eigen::Vector3d Shape::get_center_position() const {
+  return this->center_state_.get_position();
+}
+
+const Eigen::Quaterniond Shape::get_center_orientation() const {
+  return this->center_state_.get_orientation();
+}
+
+const CartesianTwist& Shape::get_center_twist() const {
+  return static_cast<const CartesianTwist&>(this->center_state_);
+}
+
+void Shape::set_center_state(const CartesianState& state) {
+  if (state.is_empty()) {
+    throw exceptions::EmptyStateException(state.get_name() + " state is empty");
+  }
+  this->center_state_ = state;
+  this->set_empty(false);
+  this->reset_timestamp();
+}
+
+void Shape::set_center_pose(const CartesianPose& pose) {
+  if (this->center_state_.get_reference_frame() != pose.get_reference_frame()) {
+    throw exceptions::IncompatibleReferenceFramesException(
+        "The shape state and the given pose are not expressed in the same reference frame");
+  }
+  this->center_state_.set_pose(pose.get_position(), pose.get_orientation());
+  this->set_empty(false);
+  this->reset_timestamp();
+}
+
+void Shape::set_center_position(const Eigen::Vector3d& position) {
+  this->center_state_.set_position(position);
+  this->set_empty(false);
+  this->reset_timestamp();
+}
+
+void Shape::set_center_orientation(const Eigen::Quaterniond& orientation) {
+  this->center_state_.set_orientation(orientation);
+  this->set_empty(false);
+  this->reset_timestamp();
+}
+
 std::string Shape::to_string() const {
   std::stringstream s;
   s << this->State::to_string();
@@ -38,4 +98,4 @@ std::ostream& operator<<(std::ostream& os, const Shape& shape) {
   os << shape.to_string();
   return os;
 }
-}
+}// namespace state_representation
